Refactor: split MangaViewModel into domain-specific extensions
1490行の巨大ファイルを8つの論理単位に分割:
- MangaViewModel.swift (288行): プロパティ、init、fetch、save
- +Migration.swift (174行): 起動時マイグレーション、dedupe
- +CRUD.swift (286行): エントリのadd/update/delete/move
- +Publisher.swift (135行): 掲載誌統合、メタデータ管理
- +Backup.swift (162行): バックアップexport/import
- +Comments.swift (76行): コメントCRUD
- +Links.swift (67行): リンクCRUD
- +ReadState.swift (289行): 既読/未読、フォーカス積読、ソフトデリート

動作変更なし。全ファイル300行以下。
